data-flow-and-pipelining/requirements.txt
apache-airflow-providers-dbt-cloud==2.3.1